Refactor: moved 'Disable', 'Rename' and 'Remove' to separate functions and separated out their tree view counterparts; Removed `GNURadioController` class
